@@ -12,18 +12,18 @@ readme = "README.md"
 documentation = "https://rob-blackbourn.github.io/jetblack-fixparser"
 
 [tool.poetry.dependencies]
-python = "^3.7"
-wheel = "^0.34.2"
-"ruamel.yaml" = "^0.16.10"
+python = "^3.8"
+# wheel = "^0.34.2"
+"ruamel.yaml" = "^0.17"
 
 [tool.poetry.dev-dependencies]
-pytest = "^4.2"
-pytest-runner = "^4.2"
-pylint = "^2.5.3"
-autopep8 = "^1.5.3"
-mypy = "^0.790"
-mkdocs = "^1.0.4"
-mkdocs-material = "^4.6.0"
+pytest = "^7.2"
+pytest-runner = "^6.0"
+pylint = "^2.17"
+autopep8 = "^2.0"
+mypy = "^1.1"
+mkdocs = "^1.4"
+mkdocs-material = "^9.1"
 jetblack-markdown = "^0.4"
 
 [build-system]
