@@ -1,13 +1,6 @@
 """Meta data for the FIX message"""
 
-from typing import Mapping, Union
-
-from .message_member import MessageMemberMetaData
-
-MessageFieldMetaDataMapping = Mapping[
-    str,
-    Union[MessageMemberMetaData, 'MessageFieldMetaDataMapping']
-]
+from .message_member import MessageFieldMetaDataMapping
 
 
 class MessageMetaData:
@@ -36,13 +29,9 @@ class MessageMetaData:
     def __str__(self) -> str:
         return (
             'MessageMetaData: '
-            'name="{name}", '
-            'msgtype="{msgtype}", '
-            'msgcat="{msgcat}"'
-        ).format(
-            name=self.name,
-            msgtype=self.msgtype.decode('ascii'),
-            msgcat=self.msgcat
+            f'name="{self.name}", '
+            f'msgtype="{self.msgtype.decode("ascii")}", '
+            f'msgcat="{self.msgcat}"'
         )
 
     __repr__ = __str__
